requirements.txt
asgiref==3.6.0
dbr==9.6.20
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
Pillow==9.5.0
Pygments==2.15.1
PyJWT==2.7.0
pypng==0.20220715.0
pytz==2023.3
pyzbar==0.1.9
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.5.0
